feat: play-routine 기능 분리
